-Main menu username field now displays the steam name. -Profile window username field now displays the steam name. -Inputfield for modifying the username locked. -Current energy amount text of the roulette screen now displays the energy that the player has. -Weigh anchor button of the waiting room now grays out afer being clicked so its clear that the user doesn't need or should press it again. -Added comments to GPProfileWindow class. -Added comments to GPLevelUpScreen class. -Added comments to GPWeeklyRewardScreen class. -Added comments to GPFriendsTaps class. -Cleaned and added comments to the GPPlayerProfile class. -Moved the code for openning chests to a new class because that same logic needs to be used from many places (store, roulette, weekly rewards, level up rewards [maybe]).
