+++ Assets/Prefabs/UI/FriendSlotUI.prefab
@@ -1724,9 +1724,9 @@ MonoBehaviour:
   m_statusIcon: {fileID: 5391126101812049882}
   m_offlineSprite: {fileID: 21300000, guid: b2b2b5de96fed40d2898dd15c17892e6, type: 3}
   m_onlineSprite: {fileID: 21300000, guid: 4ad15db1aebb749f5b1846fbbfa25de7, type: 3}
-  m_addButton: {fileID: 0}
-  m_chatButton: {fileID: 0}
-  m_blockButton: {fileID: 0}
+  m_addButton: {fileID: 2336857347621304783}
+  m_chatButton: {fileID: 1254704006904994089}
+  m_blockButton: {fileID: 6093638423634926656}
   onAddButtonClickedEvent:
     m_PersistentCalls:
       m_Calls: []
