+++ Assets/Scripts/Menus/GPChestOpeningLogic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03b9d7d4de81677429c61313ca416676
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
