+++ Assets/Scripts/Menus/GPWeeklyRewardScreen.cs
@@ -13,7 +13,7 @@ public class GPWeeklyRewardScreen : GPGUIScreen
         public int m_amount = 0;
     }
 
-
+    [Header("Reward settings")]
     public TextMeshProUGUI m_rewardCounterText;
     public List<GPWeeklyRewardBlock> m_rewardBlocks;
     [Tooltip("Prizes to give each day in order. Only the first 7 will be used (one per day).")]
@@ -26,16 +26,13 @@ public class GPWeeklyRewardScreen : GPGUIScreen
     [Header("Audio settings")]
     public AudioClip m_showScreenSFX;
     public AudioClip m_claimSFX;
-    public AudioClip m_chestOpenSFX;
 
     [Header("Misc.")]
     [SerializeField]
     private GameObject m_LoadIndicator;
 
     [Header("Chest Rewards Settings")]
-    public GPChestRewardWindow m_rewardWindow;
-    public GPGUIScreen m_rewardScreen;
-    
+    public GPChestOpeningLogic m_chestOpeningLogic;
 
     // Start is called before the first frame update
     void Start()
@@ -57,6 +54,9 @@ public class GPWeeklyRewardScreen : GPGUIScreen
         base.Hide();
     }
 
+    /// <summary>
+    /// Gives the rewards to the player, play effects and updates the claimed reward number.
+    /// </summary>
     public void ClaimReward()
     {
         GivePrize(m_rewards[m_currClaimedIdx]);
@@ -66,6 +66,9 @@ public class GPWeeklyRewardScreen : GPGUIScreen
         Hide();
     }
 
+    /// <summary>
+    /// Resets the current reward to claim to be the first one.
+    /// </summary>
     public void ResetWeekProgress()
     {
         m_currClaimedIdx = 0;
@@ -73,8 +76,9 @@ public class GPWeeklyRewardScreen : GPGUIScreen
     }
 
     /// <summary>
-    /// Displays the reward sprites, eneables check marks on the claimed ones, disiplays the number of claimed rewards
-    /// and enables the focus sprite on the current reward block.
+    /// Displays the reward sprites, eneables check marks on the claimed ones,
+    /// displays the number of claimed rewards and enables the focus sprite on
+    /// the current reward block.
     /// </summary>
     void UpdateDisplayedData()
     {
@@ -90,6 +94,10 @@ public class GPWeeklyRewardScreen : GPGUIScreen
         }
     }
 
+    /// <summary>
+    /// Gives the given weekly reward to the player.
+    /// </summary>
+    /// <param name="reward"></param>
     public async void GivePrize(WeeklyReward reward)
     {
         switch (reward.m_rewardSO.m_type)
@@ -114,7 +122,7 @@ public class GPWeeklyRewardScreen : GPGUIScreen
                     {
                         chests.Add(GPItemsDB.m_instance.m_woodenChest);
                     }
-                    OpenChestsInSequence(chests);
+                    m_chestOpeningLogic.OpenChestsInSequence(chests);
                     break;
                 }
             case GP_PRIZE_TYPE.kSilverChest:
@@ -124,7 +132,7 @@ public class GPWeeklyRewardScreen : GPGUIScreen
                     {
                         chests.Add(GPItemsDB.m_instance.m_silverChest);
                     }
-                    OpenChestsInSequence(chests);
+                    m_chestOpeningLogic.OpenChestsInSequence(chests);
                     break;
                 }
             case GP_PRIZE_TYPE.kGoldenChest:
@@ -134,7 +142,7 @@ public class GPWeeklyRewardScreen : GPGUIScreen
                     {
                         chests.Add(GPItemsDB.m_instance.m_goldenChest);
                     }
-                    OpenChestsInSequence(chests);
+                    m_chestOpeningLogic.OpenChestsInSequence(chests);
                     break;
                 }
             case GP_PRIZE_TYPE.kCrystalChest:
@@ -144,50 +152,12 @@ public class GPWeeklyRewardScreen : GPGUIScreen
                     {
                         chests.Add(GPItemsDB.m_instance.m_crystalChest);
                     }
-                    OpenChestsInSequence(chests);
+                    m_chestOpeningLogic.OpenChestsInSequence(chests);
                     break;
                 }
             default:
                 break;
         }
-    }
-
-    public void OpenChestsInSequence(List<GPStoreChestSO> chests)
-    {
-        StartCoroutine(IEOpenChestsInSecuence(chests));
-    }
-
-    IEnumerator IEOpenChestsInSecuence(List<GPStoreChestSO> chests)
-    {
-        for (int i = 0; i < chests.Count; i++)
-        {
-            OpenChest(chests[i]);
-            yield return new WaitForSeconds(3.0f);
-            if (i < chests.Count - 1) // so the last one doesn't play a sound at the end
-            {
-                TanksMP.AudioManager.Play2D(m_chestOpenSFX);
-            }
-        }
-    }
-
-    public void OpenChest(GPStoreChestSO chestDesc)
-    {
-        GPGivenRewards rewards = chestDesc.OpenChest();
-        m_rewardScreen.Show();
-        m_rewardWindow.Show();
-        m_rewardWindow.ClearContent();
-        m_rewardWindow.DisplayChestImage(chestDesc);
-        m_rewardWindow.DisplayCrewRewards(rewards.m_ships);
-        m_rewardWindow.DisplayIconRewards(rewards.m_profileIcons);
-        m_rewardWindow.DisplayDummyRewards(rewards.m_dummyParts);
-        StartCoroutine(CloseRewardWindow()); // for now close reward window after 3 seconds
-    }
-
-    IEnumerator CloseRewardWindow()
-    {
-        yield return new WaitForSeconds(3.0f);
-        m_rewardScreen.Hide();
-        m_rewardWindow.Hide();
     }
 
 }
