+++ Assets/Scripts/UI/GPUserProfileUI.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
 using TanksMP;
+using Photon.Pun;
 
 public class GPUserProfileUI : MonoBehaviour
 {
@@ -12,15 +13,12 @@ public class GPUserProfileUI : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        if (PlayerPrefs.HasKey(PrefsKeys.playerName))
-        {
-            m_inputField.text = PlayerPrefs.GetString(PrefsKeys.playerName);
-        }
+        m_inputField.text = APIManager.Instance.PlayerData.Name;
     }
 
     public void UpdateName()
     {
         PlayerPrefs.SetString(PrefsKeys.playerName, m_inputField.text);
     }
-    
+
 }
