+++ Assets/_Main/Scenes/Waiting_Room.unity
@@ -14288,6 +14288,7 @@ MonoBehaviour:
   m_timersText:
   - {fileID: 397835791}
   - {fileID: 891353810}
+  m_weighAnchorButton: {fileID: 1659599171691007784}
   m_selectTeamText: Select Team
   m_searchingForPlayersText: Searching Players
   m_matchFoundText: Match Found
